fastapi-mysql/requirements.txt
fastapi==0.95.1
uvicorn[standard]==0.22.0
mysql-connector-python==8.0.33